+++ Sources/Sentry/SentrySession.m
@@ -2,7 +2,6 @@
 #import "NSMutableDictionary+Sentry.h"
 #import "SentryCurrentDateProvider.h"
 #import "SentryDependencyContainer.h"
-#import "SentryInstallation.h"
 #import "SentryLog.h"
 #import "SentrySession+Private.h"
 
@@ -32,7 +31,7 @@ nameForSentrySessionStatus(SentrySessionStatus status)
  * Default private constructor. We don't name it init to avoid the overlap with the default init of
  * NSObject, which is not available as we specified in the header with SENTRY_NO_INIT.
  */
-- (instancetype)initDefault
+- (instancetype)initDefault:(NSString*)distinctId
 {
     if (self = [super init]) {
         _sessionId = [NSUUID UUID];
@@ -40,15 +39,15 @@ nameForSentrySessionStatus(SentrySessionStatus status)
         _status = kSentrySessionStatusOk;
         _sequence = 1;
         _errors = 0;
-        _distinctId = [SentryInstallation id];
+        _distinctId = distinctId;
     }
 
     return self;
 }
 
-- (instancetype)initWithReleaseName:(NSString *)releaseName
+- (instancetype)initWithReleaseName:(NSString *)releaseName distinctId:(NSString*)distinctId
 {
-    if (self = [self initDefault]) {
+    if (self = [self initDefault:distinctId]) {
         _init = @YES;
         _releaseName = releaseName;
     }