+++ Sources/SentryCrash/Recording/SentryCrash.h
@@ -28,6 +28,7 @@
 #import "SentryCrashMonitorType.h"
 #import "SentryCrashReportFilter.h"
 #import "SentryCrashReportWriter.h"
+#import "SentryDefines.h"
 
 typedef enum {
     SentryCrashDemangleLanguageNone = 0,
@@ -55,9 +56,13 @@ static NSString *const SENTRYCRASH_REPORT_ATTACHMENTS_ITEM = @"attachments";
 @interface SentryCrash : NSObject
 
 #pragma mark - Configuration -
+SENTRY_NO_INIT
 
 /** Init SentryCrash instance with custom base path. */
-- (id)initWithBasePath:(NSString *)basePath;
+- (instancetype)initWithBasePath:(NSString *)basePath NS_DESIGNATED_INITIALIZER;
+
+/** Cache directory base path. */
+@property (nonatomic, readwrite, retain) NSString *basePath;
 
 /** A dictionary containing any info you'd like to appear in crash reports. Must
  * contain only JSON-safe data: NSString for keys, and NSDictionary, NSArray,