+++ Tests/SentryTests/SentryCrash/SentryCrash+Test.h
@@ -1,0 +1,5 @@
+#import "SentryCrash.h"
+
+@interface SentryCrash (Test)
+- (NSString *)getBundleName;
+@end