+++ Tests/SentryTests/State/SentryInstallation+Test.h
@@ -1,0 +1,5 @@
+#import "SentryInstallation.h"
+
+@interface SentryInstallation (Test)
+@property(class, nonatomic, readonly) NSMutableDictionary<NSString*, NSString*>* installationStringsByCacheDirectoryPaths;
+@end